sql: refactor execution detail utility methods
This change is a refactor only commit to move some of
the utility methods used to read, write and list execution
detail files to `pkg/jobs`. This allows us to use these
methods without pulling in a dependancy to the sql package.

Release note: None
