+++ pkg/jobs/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "adopt.go",
         "config.go",
         "errors.go",
+        "execution_detail_utils.go",
         "executor_impl.go",
         "helpers.go",
         "job_info_storage.go",
@@ -31,6 +32,7 @@ go_library(
         "//pkg/base",
         "//pkg/clusterversion",
         "//pkg/jobs/jobspb",
+        "//pkg/jobs/jobsprofiler/profilerconstants",
         "//pkg/kv",
         "//pkg/kv/kvpb",
         "//pkg/multitenant",
@@ -75,6 +77,7 @@ go_library(
         "@com_github_cockroachdb_redact//:redact",
         "@com_github_gogo_protobuf//jsonpb",
         "@com_github_gogo_protobuf//types",
+        "@com_github_klauspost_compress//gzip",
         "@com_github_prometheus_client_model//go",
         "@com_github_robfig_cron_v3//:cron",
         "@io_opentelemetry_go_otel//attribute",
