+++ pkg/sql/BUILD.bazel
@@ -141,8 +141,7 @@ go_library(
         "job_exec_context.go",
         "job_exec_context_test_util.go",
         "jobs_collection.go",
-        "jobs_execution_details.go",
-        "jobs_profiler_bundle.go",
+        "jobs_profiler_execution_details.go",
         "join.go",
         "join_predicate.go",
         "join_token.go",
@@ -577,7 +576,6 @@ go_library(
         "@com_github_dustin_go_humanize//:go-humanize",
         "@com_github_gogo_protobuf//proto",
         "@com_github_gogo_protobuf//types",
-        "@com_github_klauspost_compress//gzip",
         "@com_github_lib_pq//:pq",
         "@com_github_lib_pq//oid",
         "@com_github_prometheus_client_model//go",
@@ -649,8 +647,7 @@ go_test(
         "indexbackfiller_test.go",
         "instrumentation_test.go",
         "internal_test.go",
-        "jobs_execution_details_test.go",
-        "jobs_profiler_bundle_test.go",
+        "jobs_profiler_execution_details_test.go",
         "join_token_test.go",
         "main_test.go",
         "materialized_view_test.go",
@@ -732,7 +729,6 @@ go_test(
     deps = [
         "//pkg/base",
         "//pkg/build/bazel",
-        "//pkg/cloud/impl:cloudimpl",
         "//pkg/clusterversion",
         "//pkg/col/coldata",
         "//pkg/config",
