+++ pkg/sql/jobs_profiler_execution_details.go
@@ -13,16 +13,22 @@ package sql
 import (
 	"context"
 	gojson "encoding/json"
+	"fmt"
 	"net/url"
 	"strconv"
 
 	"github.com/cockroachdb/cockroach/pkg/base"
+	"github.com/cockroachdb/cockroach/pkg/clusterversion"
 	"github.com/cockroachdb/cockroach/pkg/jobs"
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobsprofiler/profilerconstants"
+	"github.com/cockroachdb/cockroach/pkg/server/serverpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfrapb"
 	"github.com/cockroachdb/cockroach/pkg/sql/isql"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/eval"
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
+	"github.com/cockroachdb/cockroach/pkg/sql/sessiondata"
+	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
 	"github.com/cockroachdb/errors"
 )
@@ -165,4 +171,80 @@ func constructBackupExecutionDetails(
 	}
 	j, err := gojson.Marshal(executionDetails)
 	return j, err
+}
+
+// RequestExecutionDetailFiles implements the JobProfiler interface.
+func (p *planner) RequestExecutionDetailFiles(ctx context.Context, jobID jobspb.JobID) error {
+	execCfg := p.ExecCfg()
+	if !execCfg.Settings.Version.IsActive(ctx, clusterversion.V23_1) {
+		return errors.Newf("execution details can only be requested on a cluster with version >= %s",
+			clusterversion.V23_1.String())
+	}
+
+	e := makeJobProfilerExecutionDetailsBuilder(execCfg.SQLStatusServer, execCfg.InternalDB, jobID)
+	// TODO(adityamaru): When we start collecting more information we can consider
+	// parallelize the collection of the various pieces.
+	e.addDistSQLDiagram(ctx)
+	e.addLabelledGoroutines(ctx)
+
+	return nil
+}
+
+// executionDetailsBuilder can be used to read and write execution details corresponding
+// to a job.
+type executionDetailsBuilder struct {
+	srv   serverpb.SQLStatusServer
+	db    isql.DB
+	jobID jobspb.JobID
+}
+
+// makeJobProfilerExecutionDetailsBuilder returns an instance of an executionDetailsBuilder.
+func makeJobProfilerExecutionDetailsBuilder(
+	srv serverpb.SQLStatusServer, db isql.DB, jobID jobspb.JobID,
+) executionDetailsBuilder {
+	e := executionDetailsBuilder{
+		srv: srv, db: db, jobID: jobID,
+	}
+	return e
+}
+
+// addLabelledGoroutines collects and persists goroutines from all nodes in the
+// cluster that have a pprof label tying it to the job whose execution details
+// are being collected.
+func (e *executionDetailsBuilder) addLabelledGoroutines(ctx context.Context) {
+	profileRequest := serverpb.ProfileRequest{
+		NodeId:      "all",
+		Type:        serverpb.ProfileRequest_GOROUTINE,
+		Labels:      true,
+		LabelFilter: fmt.Sprintf("%d", e.jobID),
+	}
+	resp, err := e.srv.Profile(ctx, &profileRequest)
+	if err != nil {
+		log.Errorf(ctx, "failed to collect goroutines for job %d: %+v", e.jobID, err.Error())
+		return
+	}
+	filename := fmt.Sprintf("goroutines.%s.txt", timeutil.Now().Format("20060102_150405.00"))
+	if err := jobs.WriteExecutionDetailFile(ctx, filename, resp.Data, e.db, e.jobID); err != nil {
+		log.Errorf(ctx, "failed to write goroutine for job %d: %+v", e.jobID, err.Error())
+	}
+}
+
+// addDistSQLDiagram generates and persists a `distsql.<timestamp>.html` file.
+func (e *executionDetailsBuilder) addDistSQLDiagram(ctx context.Context) {
+	query := `SELECT plan_diagram FROM [SHOW JOB $1 WITH EXECUTION DETAILS]`
+	row, err := e.db.Executor().QueryRowEx(ctx, "profiler-bundler-add-diagram", nil, /* txn */
+		sessiondata.NoSessionDataOverride, query, e.jobID)
+	if err != nil {
+		log.Errorf(ctx, "failed to write DistSQL diagram for job %d: %+v", e.jobID, err.Error())
+		return
+	}
+	if row[0] != tree.DNull {
+		dspDiagramURL := string(tree.MustBeDString(row[0]))
+		filename := fmt.Sprintf("distsql.%s.html", timeutil.Now().Format("20060102_150405.00"))
+		if err := jobs.WriteExecutionDetailFile(ctx, filename,
+			[]byte(fmt.Sprintf(`<meta http-equiv="Refresh" content="0; url=%s">`, dspDiagramURL)),
+			e.db, e.jobID); err != nil {
+			log.Errorf(ctx, "failed to write DistSQL diagram for job %d: %+v", e.jobID, err.Error())
+		}
+	}
 }
